Implement withdraw amount

+++ src/main/java/com/priceminister/account/IllegalBalanceException.java
@@ -1,13 +1,15 @@
 package com.priceminister.account;
 
 
+import java.math.BigDecimal;
+
 public class IllegalBalanceException extends Exception {
     
     private static final long serialVersionUID = -9204191749972551939L;
     
-	private Double balance;
+	private BigDecimal balance;
     
-    public IllegalBalanceException(Double illegalBalance) {
+    public IllegalBalanceException(BigDecimal illegalBalance) {
         balance = illegalBalance;
     }
